+++ packages/sh-design/src/components/index.ts
@@ -1,4 +1,3 @@
-export * from './copy-button'
 export * from './lazy-image'
 export * from './seamless-scroll'
 export * from './waterfall'